Fixed bug in the code!
I moved the attempts++; from above the scope of the userInput validation range. I think that with it above the scope it was counting the out of range input as an attempt. I moved it below the validation range to the next block. it seems to have solved the problem.